Add files via upload
TP4_partie1: This notebook uses linear regression to model a dataset with scikit-learn. It initializes a linear model, fits it to data points, and outputs the slope and intercept. The model's predictions are visualized in a scatter plot with a best-fit line, and the mean squared error is calculated to evaluate performance.

TP4_partie2: This notebook builds on the previous one by applying polynomial regression. After transforming the data with polynomial features, it fits an 8th-degree polynomial model, makes predictions, and visualizes the results in a plot with the curve passing through the data points, showing a more flexible fit compared to the linear model. ​